Update source/js/main.js
Update source/js/main.js

fixes

add local

design feedback

fix linter

done

donate fix

moved to seperate file

update

lint fix

percy fix

oops
