+++ network-api/networkapi/templates/partials/banner.html
@@ -1,16 +1,16 @@
 {% load settings_value i18n static %}
 
 {% with wrapper_class="tw-py-7" %}
-<div class="donate-banner tw-bg-black dark-theme {{wrapper_class}} print:tw-hidden tw-w-full">
-    <a href="#" class="p-4 tw-box-content banner-close" aria-label="Close Banner">
-        <img src="{% static "_images/x.svg" %}" class="tw-w-4 tw-h-4 p-4 tw-box-content tw-absolute tw-right-0 tw-top-0 tw-cursor-pointer">
+<div class="donate-banner tw-bg-black dark-theme {{wrapper_class}} print:tw-hidden tw-w-full tw-hidden">
+    <a href="#" class="p-4 tw-box-content banner-close tw-cursor-pointer tw-absolute tw-right-0 tw-top-0 tw-group" aria-label="{% trans "Close Banner" %}">
+        <img src="{% static "_images/x.svg" %}" class="tw-w-4 tw-h-4 tw-transition-transform tw-duration-100 tw-ease-in-out tw-transform group-hover:tw-scale-110">
     </a>
 
     <div class="tw-container">
         <div class="tw-row">
 
             {% block left_banner_column %}
-            <div class="medium:tw-w-5/12 large:tw-w-6/12 tw-px-4">
+            <div class="medium:tw-w-6/12 large:tw-w-6/12 tw-px-4">
                 <p class="h2-heading tw-mb-0">
                     {% trans "Help Mozilla fight for a better internet this Giving Tuesday" %}
                 </p>
@@ -35,7 +35,7 @@
                 </div>
                 <div class="tw-row">
                     <div class="tw-w-full tw-px-4">
-                        <a class="btn btn-banner tw-w-full" href="https://donate.mozilla.org/?utm_source=foundation.mozilla.org&utm_medium=referral&utm_campaign=fmonav&utm_content=header" target="_blank" rel="noopener noreferrer">{% trans "Donate" %}</a>
+                        <a class="btn btn-banner tw-w-full" href="https://donate.mozilla.org/?utm_source=foundation.mozilla.org&utm_medium=referral&utm_campaign=fmonav&utm_content=header" target="_blank" rel="noopener noreferrer">{% trans "Support Mozilla" %}</a>
                     </div>
                 </div>
             </div>
