+++ source/sass/mofo-bootstrap/_colors.scss
@@ -51,6 +51,11 @@ $youtube-regrets-dark-gray: #696969;
 $dear-internet-lilac: #d3d5fc;
 $dear-internet-pink: #fff0f1;
 
+// Banner Button Palatte
+$banner-default-color: #00ffff;
+$banner-hover-color: #00c2c2;
+$banner-active-color: #0d6c6c;
+
 // Map of all our brand colours defined above
 
 $colors: (
@@ -78,4 +83,7 @@ $colors: (
   "youtube-regrets-light-red": $youtube-regrets-light-red,
   "dear-internet-lilac": $dear-internet-lilac,
   "dear-internet-pink": $dear-internet-pink,
+  "banner-default-color": $banner-default-color,
+  "banner-hover-color": $banner-hover-color,
+  "banner-active-color": $banner-active-color,
 );
